requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
fake-useragent==0.1.13
idna==3.4
lxml==4.9.1
pyTelegramBotAPI==4.7.1
python-dotenv==0.21.0
requests==2.28.1
soupsieve==2.3.2.post1
urllib3==1.26.12